+++ src/test/persistence/TestJson.java
@@ -5,8 +5,7 @@ import model.Investment;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestJson {
-    protected void checkInvestment(String type, String name, int amount, int expRet, String date, 
-    Investment i) {
+    protected void checkInvestment(String type, String name, double amount, double expRet, String date, Investment i) {
         assertEquals(type,i.getType());
         assertEquals(name,i.getName());
         assertEquals(amount,i.getAmount());
